mvp-backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
PyJWT==2.9.0
python-dotenv==1.0.1
